# openapi.yaml
openapi: 3.0.3
info:
  title: Task Tracker API
  description: RESTful API for managing projects, users, and tasks.
  version: 1.0.0
  contact:
    name: Sagar Agrawal

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/info:
    get:
      summary: Application info
      operationId: getAppInfo
      tags:
        - Info
      responses:
        '200':
          description: Application name and version
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: Task Tracker API
                  version:
                    type: string
                    example: 1.0.0

  /api/projects:
    post:
      summary: Create a new project
      operationId: createProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Project name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List all projects
      operationId: getAllProjects
      tags:
        - Projects
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedProjectResponse'

  /api/projects/{id}:
    get:
      summary: Get project by ID
      operationId: getProjectById
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a project
      operationId: updateProject
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found

    delete:
      summary: Delete a project
      operationId: deleteProject
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Project deleted
        '404':
          description: Project not found

  /api/projects/{id}/tasks:
    get:
      summary: Get all tasks for a project
      operationId: getTasksByProject
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: List of tasks under the project
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Project not found

  /api/users:
    post:
      summary: Create a user
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
        '409':
          description: Email already exists

    get:
      summary: List all users
      operationId: getAllUsers
      tags:
        - Users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUserResponse'

  /api/users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found

    put:
      summary: Update a user
      operationId: updateUser
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found

    delete:
      summary: Delete a user
      operationId: deleteUser
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found

  /api/tasks:
    post:
      summary: Create a task under a project
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
        '404':
          description: Referenced project or user not found

    get:
      summary: List all tasks (filter by status or project)
      operationId: getAllTasks
      tags:
        - Tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [TODO, IN_PROGRESS, DONE]
        - name: projectId
          in: query
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated list of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedTaskResponse'

  /api/tasks/{id}:
    get:
      summary: Get task by ID
      operationId: getTaskById
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Task found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found

    put:
      summary: Update task details or status
      operationId: updateTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task or referenced entity not found

    delete:
      summary: Delete a task
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Task deleted
        '404':
          description: Task not found

components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    ProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Backend API
        description:
          type: string
          example: Core backend service for the platform

    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time

    UserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          example: Alice Johnson
        email:
          type: string
          format: email
          example: alice@example.com
        role:
          type: string
          example: DEVELOPER

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        role:
          type: string

    TaskRequest:
      type: object
      required:
        - title
        - projectId
      properties:
        title:
          type: string
          example: Implement user login
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        projectId:
          type: integer
          format: int64
          example: 1
        assigneeId:
          type: integer
          format: int64
          example: 1
        dueDate:
          type: string
          format: date
          example: '2026-06-15'

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        projectId:
          type: integer
          format: int64
        projectName:
          type: string
        assigneeId:
          type: integer
          format: int64
        assigneeName:
          type: string
        dueDate:
          type: string
          format: date

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        errors:
          type: array
          items:
            type: string

    PagedProjectResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    PagedUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    PagedTaskResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer